fix(config): Fix up bad values in option lists
Sometimes it appears that people can end up with bad values in their config for option
lists - e.g. one guy somehow had the wrong thing for his rollOptions. This was breaking the
UI previously, so change it to replace these bad values with the first value in the list.